+++ aider/coders/base_coder.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 
+import base64
 import hashlib
 import json
 import locale
@@ -13,9 +14,11 @@ import threading
 import time
 import traceback
 from collections import defaultdict
+from dataclasses import dataclass, field
 from datetime import datetime
 from json.decoder import JSONDecodeError
 from pathlib import Path
+from typing import List
 
 import git
 from rich.console import Console, Text
@@ -48,6 +51,57 @@ def wrap_fence(name):
     return f"<{name}>", f"</{name}>"
 
 
+@dataclass
+class ChatChunks:
+    system: List = field(default_factory=list)
+    examples: List = field(default_factory=list)
+    done: List = field(default_factory=list)
+    repo: List = field(default_factory=list)
+    readonly_files: List = field(default_factory=list)
+    chat_files: List = field(default_factory=list)
+    cur: List = field(default_factory=list)
+    reminder: List = field(default_factory=list)
+
+    def all_messages(self):
+        return (
+            self.system
+            + self.examples
+            + self.repo
+            + self.readonly_files
+            + self.done
+            + self.chat_files
+            + self.cur
+            + self.reminder
+        )
+
+    def add_cache_control_headers(self):
+        if self.examples:
+            self.add_cache_control(self.examples)
+        else:
+            self.add_cache_control(self.system)
+
+        if self.readonly_files:
+            self.add_cache_control(self.readonly_files)
+        else:
+            self.add_cache_control(self.repo)
+
+        self.add_cache_control(self.chat_files)
+
+    def add_cache_control(self, messages):
+        if not messages:
+            return
+
+        content = messages[-1]["content"]
+        if type(content) is str:
+            content = dict(
+                type="text",
+                text=content,
+            )
+        content["cache_control"] = {"type": "ephemeral"}
+
+        messages[-1]["content"] = [content]
+
+
 class Coder:
     abs_fnames = None
     abs_read_only_fnames = None
@@ -73,6 +127,9 @@ class Coder:
     multi_response_content = ""
     partial_response_content = ""
     commit_before_message = []
+    message_cost = 0.0
+    message_tokens_sent = 0
+    message_tokens_received = 0
 
     @classmethod
     def create(
@@ -149,7 +206,10 @@ class Coder:
         main_model = self.main_model
         weak_model = main_model.weak_model
         prefix = "Model:"
-        output = f" {main_model.name} with {self.edit_format} edit format"
+        output = f" {main_model.name} with"
+        if main_model.info.get("supports_assistant_prefill"):
+            output += " ♾️"
+        output += f" {self.edit_format} edit format"
         if weak_model is not main_model:
             prefix = "Models:"
             output += f", weak model {weak_model.name}"
@@ -172,7 +232,8 @@ class Coder:
         if self.repo_map:
             map_tokens = self.repo_map.max_map_tokens
             if map_tokens > 0:
-                lines.append(f"Repo-map: using {map_tokens} tokens")
+                refresh = self.repo_map.refresh
+                lines.append(f"Repo-map: using {map_tokens} tokens, {refresh} refresh")
                 max_map_tokens = 2048
                 if map_tokens > max_map_tokens:
                     lines.append(
@@ -222,11 +283,14 @@ class Coder:
         commands=None,
         summarizer=None,
         total_cost=0.0,
+        map_refresh="auto",
+        cache_prompts=False,
     ):
         self.commit_before_message = []
         self.aider_commit_hashes = set()
         self.rejected_urls = set()
         self.abs_root_path_cache = {}
+        self.cache_prompts = cache_prompts
 
         if not fnames:
             fnames = []
@@ -351,6 +415,7 @@ class Coder:
                 self.verbose,
                 max_inp_tokens,
                 map_mul_no_files=map_mul_no_files,
+                refresh=map_refresh,
             )
 
         self.summarizer = summarizer or ChatSummary(
@@ -588,40 +653,24 @@ class Coder:
 
         return repo_content
 
-    def get_files_messages(self):
-        files_messages = []
-
+    def get_repo_messages(self):
+        repo_messages = []
         repo_content = self.get_repo_map()
         if repo_content:
-            files_messages += [
+            repo_messages += [
                 dict(role="user", content=repo_content),
                 dict(
                     role="assistant",
                     content="Ok, I won't try and edit those files without asking first.",
                 ),
             ]
+        return repo_messages
 
-        if self.abs_fnames:
-            files_content = self.gpt_prompts.files_content_prefix
-            files_content += self.get_files_content()
-            files_reply = "Ok, any changes I propose will be to those files."
-        elif repo_content and self.gpt_prompts.files_no_full_files_with_repo_map:
-            files_content = self.gpt_prompts.files_no_full_files_with_repo_map
-            files_reply = self.gpt_prompts.files_no_full_files_with_repo_map_reply
-        else:
-            files_content = self.gpt_prompts.files_no_full_files
-            files_reply = "Ok."
-
-        images_message = self.get_images_message()
-        if images_message is not None:
-            files_messages += [
-                images_message,
-                dict(role="assistant", content="Ok."),
-            ]
-
+    def get_readonly_files_messages(self):
+        readonly_messages = []
         read_only_content = self.get_read_only_files_content()
         if read_only_content:
-            files_messages += [
+            readonly_messages += [
                 dict(
                     role="user", content=self.gpt_prompts.read_only_files_prefix + read_only_content
                 ),
@@ -630,14 +679,35 @@ class Coder:
                     content="Ok, I will use these files as references.",
                 ),
             ]
+        return readonly_messages
+
+    def get_chat_files_messages(self):
+        chat_files_messages = []
+        if self.abs_fnames:
+            files_content = self.gpt_prompts.files_content_prefix
+            files_content += self.get_files_content()
+            files_reply = "Ok, any changes I propose will be to those files."
+        elif self.get_repo_map() and self.gpt_prompts.files_no_full_files_with_repo_map:
+            files_content = self.gpt_prompts.files_no_full_files_with_repo_map
+            files_reply = self.gpt_prompts.files_no_full_files_with_repo_map_reply
+        else:
+            files_content = self.gpt_prompts.files_no_full_files
+            files_reply = "Ok."
 
         if files_content:
-            files_messages += [
+            chat_files_messages += [
                 dict(role="user", content=files_content),
                 dict(role="assistant", content=files_reply),
             ]
 
-        return files_messages
+        images_message = self.get_images_message()
+        if images_message is not None:
+            chat_files_messages += [
+                images_message,
+                dict(role="assistant", content="Ok."),
+            ]
+
+        return chat_files_messages
 
     def get_images_message(self):
         if not self.main_model.accepts_images:
@@ -646,9 +716,11 @@ class Coder:
         image_messages = []
         for fname, content in self.get_abs_fnames_content():
             if is_image_file(fname):
+                with open(fname, "rb") as image_file:
+                    encoded_string = base64.b64encode(image_file.read()).decode("utf-8")
                 mime_type, _ = mimetypes.guess_type(fname)
                 if mime_type and mime_type.startswith("image/"):
-                    image_url = f"data:{mime_type};base64,{content}"
+                    image_url = f"data:{mime_type};base64,{encoded_string}"
                     rel_fname = self.get_rel_fname(fname)
                     image_messages += [
                         {"type": "text", "text": f"Image file: {rel_fname}"},
@@ -840,8 +912,12 @@ class Coder:
         if user_lang:
             platform_text += f"- Language: {user_lang}\n"
 
-        dt = datetime.now().astimezone().strftime("%Y-%m-%dT%H:%M:%S%z")
-        platform_text += f"- Current date/time: {dt}"
+        if self.cache_prompts:
+            dt = datetime.now().astimezone().strftime("%Y-%m-%d")
+            platform_text += f"- Current date: {dt}"
+        else:
+            dt = datetime.now().astimezone().strftime("%Y-%m-%dT%H:%M:%S%z")
+            platform_text += f"- Current date/time: {dt}"
 
         prompt = prompt.format(
             fence=self.fence,
@@ -850,7 +926,7 @@ class Coder:
         )
         return prompt
 
-    def format_messages(self):
+    def format_chat_chunks(self):
         self.choose_fence()
         main_sys = self.fmt_system_prompt(self.gpt_prompts.main_system)
 
@@ -886,15 +962,19 @@ class Coder:
         if self.gpt_prompts.system_reminder:
             main_sys += "\n" + self.fmt_system_prompt(self.gpt_prompts.system_reminder)
 
-        messages = [
+        chunks = ChatChunks()
+
+        chunks.system = [
             dict(role="system", content=main_sys),
         ]
-        messages += example_messages
+        chunks.examples = example_messages
 
         self.summarize_end()
-        messages += self.done_messages
+        chunks.done = self.done_messages
 
-        messages += self.get_files_messages()
+        chunks.repo = self.get_repo_messages()
+        chunks.readonly_files = self.get_readonly_files_messages()
+        chunks.chat_files = self.get_chat_files_messages()
 
         if self.gpt_prompts.system_reminder:
             reminder_message = [
@@ -905,10 +985,13 @@ class Coder:
         else:
             reminder_message = []
 
+        chunks.cur = list(self.cur_messages)
+        chunks.reminder = []
+
         # TODO review impact of token count on image messages
-        messages_tokens = self.main_model.token_count(messages)
+        messages_tokens = self.main_model.token_count(chunks.all_messages())
         reminder_tokens = self.main_model.token_count(reminder_message)
-        cur_tokens = self.main_model.token_count(self.cur_messages)
+        cur_tokens = self.main_model.token_count(chunks.cur)
 
         if None not in (messages_tokens, reminder_tokens, cur_tokens):
             total_tokens = messages_tokens + reminder_tokens + cur_tokens
@@ -916,9 +999,7 @@ class Coder:
             # add the reminder anyway
             total_tokens = 0
 
-        messages += self.cur_messages
-
-        final = messages[-1]
+        final = chunks.cur[-1]
 
         max_input_tokens = self.main_model.info.get("max_input_tokens") or 0
         # Add the reminder prompt if we still have room to include it.
@@ -928,7 +1009,7 @@ class Coder:
             and self.gpt_prompts.system_reminder
         ):
             if self.main_model.reminder_as_sys_msg:
-                messages += reminder_message
+                chunks.reminder = reminder_message
             elif final["role"] == "user":
                 # stuff it into the user message
                 new_content = (
@@ -936,9 +1017,17 @@ class Coder:
                     + "\n\n"
                     + self.fmt_system_prompt(self.gpt_prompts.system_reminder)
                 )
-                messages[-1] = dict(role=final["role"], content=new_content)
+                chunks.cur[-1] = dict(role=final["role"], content=new_content)
 
-        return messages
+        return chunks
+
+    def format_messages(self):
+        chunks = self.format_chat_chunks()
+        if self.cache_prompts and self.main_model.cache_control:
+            chunks.add_cache_control_headers()
+
+        msgs = chunks.all_messages()
+        return msgs
 
     def send_message(self, inp):
         self.aider_edited_files = None
@@ -989,7 +1078,7 @@ class Coder:
                     return
                 except FinishReasonLength:
                     # We hit the output limit!
-                    if not self.main_model.can_prefill:
+                    if not self.main_model.info.get("supports_assistant_prefill"):
                         exhausted = True
                         break
 
@@ -998,10 +1087,13 @@ class Coder:
                     if messages[-1]["role"] == "assistant":
                         messages[-1]["content"] = self.multi_response_content
                     else:
-                        messages.append(dict(role="assistant", content=self.multi_response_content))
+                        messages.append(
+                            dict(role="assistant", content=self.multi_response_content, prefix=True)
+                        )
                 except Exception as err:
                     self.io.tool_error(f"Unexpected error: {err}")
-                    traceback.print_exc()
+                    lines = traceback.format_exception(type(err), err, err.__traceback__)
+                    self.io.tool_error("".join(lines))
                     return
         finally:
             if self.mdstream:
@@ -1013,8 +1105,7 @@ class Coder:
 
         self.io.tool_output()
 
-        if self.usage_report:
-            self.io.tool_output(self.usage_report)
+        self.show_usage_report()
 
         if exhausted:
             self.show_exhausted_error()
@@ -1024,7 +1115,7 @@ class Coder:
         if self.partial_response_function_call:
             args = self.parse_partial_args()
             if args:
-                content = args["explanation"]
+                content = args.get("explanation") or ""
             else:
                 content = ""
         elif self.partial_response_content:
@@ -1237,7 +1328,7 @@ class Coder:
 
         self.io.log_llm_history("TO LLM", format_messages(messages))
 
-        interrupted = False
+        completion = None
         try:
             hash_object, completion = send_completion(
                 model.name,
@@ -1254,9 +1345,9 @@ class Coder:
                 yield from self.show_send_output_stream(completion)
             else:
                 self.show_send_output(completion)
-        except KeyboardInterrupt:
+        except KeyboardInterrupt as kbi:
             self.keyboard_interrupt()
-            interrupted = True
+            raise kbi
         finally:
             self.io.log_llm_history(
                 "LLM RESPONSE",
@@ -1271,10 +1362,7 @@ class Coder:
                 if args:
                     self.io.ai_output(json.dumps(args, indent=4))
 
-        if interrupted:
-            raise KeyboardInterrupt
-
-        self.calculate_and_show_tokens_and_cost(messages, completion)
+            self.calculate_and_show_tokens_and_cost(messages, completion)
 
     def show_send_output(self, completion):
         if self.verbose:
@@ -1287,17 +1375,20 @@ class Coder:
         show_func_err = None
         show_content_err = None
         try:
-            self.partial_response_function_call = completion.choices[0].message.function_call
+            if completion.choices[0].message.tool_calls:
+                self.partial_response_function_call = (
+                    completion.choices[0].message.tool_calls[0].function
+                )
         except AttributeError as func_err:
             show_func_err = func_err
 
         try:
-            self.partial_response_content = completion.choices[0].message.content
+            self.partial_response_content = completion.choices[0].message.content or ""
         except AttributeError as content_err:
             show_content_err = content_err
 
         resp_hash = dict(
-            function_call=self.partial_response_function_call,
+            function_call=str(self.partial_response_function_call),
             content=self.partial_response_content,
         )
         resp_hash = hashlib.sha1(json.dumps(resp_hash, sort_keys=True).encode())
@@ -1377,22 +1468,39 @@ class Coder:
     def calculate_and_show_tokens_and_cost(self, messages, completion=None):
         prompt_tokens = 0
         completion_tokens = 0
+        cached_tokens = 0
         cost = 0
 
         if completion and hasattr(completion, "usage") and completion.usage is not None:
             prompt_tokens = completion.usage.prompt_tokens
             completion_tokens = completion.usage.completion_tokens
+            cached_tokens = getattr(completion.usage, "prompt_cache_hit_tokens", 0) or getattr(
+                completion.usage, "cache_read_input_tokens", 0
+            )
         else:
             prompt_tokens = self.main_model.token_count(messages)
             completion_tokens = self.main_model.token_count(self.partial_response_content)
 
-        self.usage_report = f"Tokens: {prompt_tokens:,} sent, {completion_tokens:,} received."
+        self.message_tokens_sent += prompt_tokens
+        self.message_tokens_received += completion_tokens
+
+        if cached_tokens:
+            tokens_report = (
+                f"Tokens: {self.message_tokens_sent:,} sent, {cached_tokens:,} cached, "
+                f"{self.message_tokens_received:,} received."
+            )
+        else:
+            tokens_report = (
+                f"Tokens: {self.message_tokens_sent:,} sent,"
+                f" {self.message_tokens_received:,} received."
+            )
 
         if self.main_model.info.get("input_cost_per_token"):
             cost += prompt_tokens * self.main_model.info.get("input_cost_per_token")
             if self.main_model.info.get("output_cost_per_token"):
                 cost += completion_tokens * self.main_model.info.get("output_cost_per_token")
             self.total_cost += cost
+            self.message_cost += cost
 
             def format_cost(value):
                 if value == 0:
@@ -1403,13 +1511,24 @@ class Coder:
                 else:
                     return f"{value:.{max(2, 2 - int(math.log10(magnitude)))}f}"
 
-            self.usage_report += (
-                f" Cost: ${format_cost(cost)} request, ${format_cost(self.total_cost)} session."
+            cost_report = (
+                f" Cost: ${format_cost(self.message_cost)} message,"
+                f" ${format_cost(self.total_cost)} session."
             )
+            self.usage_report = tokens_report + cost_report
+        else:
+            self.usage_report = tokens_report
+
+    def show_usage_report(self):
+        if self.usage_report:
+            self.io.tool_output(self.usage_report)
+            self.message_cost = 0.0
+            self.message_tokens_sent = 0
+            self.message_tokens_received = 0
 
     def get_multi_response_content(self, final=False):
-        cur = self.multi_response_content
-        new = self.partial_response_content
+        cur = self.multi_response_content or ""
+        new = self.partial_response_content or ""
 
         if new.rstrip() != new and not final:
             new = new.rstrip()
@@ -1451,7 +1570,10 @@ class Coder:
         return max(path.stat().st_mtime for path in files)
 
     def get_addable_relative_files(self):
-        return set(self.get_all_relative_files()) - set(self.get_inchat_relative_files())
+        all_files = set(self.get_all_relative_files())
+        inchat_files = set(self.get_inchat_relative_files())
+        read_only_files = set(self.get_rel_fname(fname) for fname in self.abs_read_only_fnames)
+        return all_files - inchat_files - read_only_files
 
     def check_for_dirty_commit(self, path):
         if not self.repo: